refactor: rename into_test_server_core to into_test_server_thread
